% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC.R
\name{TraitScan.MC}
\alias{TraitScan.MC}
\title{Monte Carlo Simulation}
\usage{
TraitScan.MC(
  nsim = 10000,
  p,
  alpha_grid = sort(exp(seq(log(0.05), log(0.05/p), length.out = p * 2)))
)
}
\arguments{
\item{nsim}{int, number of MC iterations.}

\item{p}{int, number of traits.}

\item{alpha_grid}{numeric vector less than 1: the alpha grid for p value thresholds. Default as a geometric sequence from the Bonferroni-significant threshold to 0.05.}
}
\value{
A list with the null distributions of TC priority function, TC statistic, and combimed test statistic
}
\description{
Monte Carlo Simulation for the null distributions of TC and combined tests
}
\note{
Use this function to simulate the null distributions before run tests \code{TraitScan.main} or \code{TraitScan.individual.level}
}
\examples{

p = 10    # number of traits
alpha_grid = sort(exp(seq(log(0.05), log(0.05/p), length.out=200)))    # alpha grid in HC statistic
nsim = 1e4    # number of MC iterations
null_dist <- TraitScan.MC(nsim = nsim, p = p, alpha_grid = alpha_grid)
}
\references{
Cao, Rui, et al. "Subset scanning for multi-trait analysis using GWAS summary statistics." Bioinformatics (2024): btad777.
}
\author{
Rui Cao, Tianzhong Yang
}
