% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_test.R
\name{TraitScan.summary.level}
\alias{TraitScan.summary.level}
\title{TraitScan Test Using Summary-level GWAS Data}
\usage{
TraitScan.summary.level(
  beta_hat,
  se_hat,
  null_cov,
  alpha_grid,
  null_dist,
  tol = 0.01
)
}
\arguments{
\item{beta_hat}{a numeric vector, estimated regression coefficients of traits.}

\item{se_hat}{a numeric vector, estimated standard errors of beta_hat.}

\item{null_cov}{a \code{p} by \code{p} positive definite matrix, the covariance matrix of z scores under the null. Can be estimated from null SNPs or LD score regression.}

\item{alpha_grid}{the alpha grid for p value thresholds. Should be the same as \code{alpha_grid} when calling \code{TraitScan.MC}}

\item{null_dist}{the output of function \code{TraitScan.MC}.}

\item{tol}{When \code{null_cov} is rank-deficient, it is recommended that the eigenvalues smaller than \code{tol} are filtered out. Additional sensitivity analysis may also be carried on \code{null_cov} to ensure the output is replicable.}
}
\value{
A list with the selected subset of traits
\item{trait.pval.decor}{The decorrelated GWAS p-values for each trait.}
\item{select.traits.minp}{The selected traits by minp strategy of HC and TC. The value represented the order of selected traits in \code{beta_hat}.}
\item{select.traits.HC}{The selected traits by HC.}
\item{select.traits.TC}{The selected traits by TC.}
\item{tests.minp}{Which test has a smaller p value, either HC or TC.}
\item{p.value}{The empirical p-value of combined test from Monte Carlo}
}
\description{
Run HC and TC tests on summary-level GWAS data and combine tests by minp strategy.
}
\examples{

set.seed(1)
p = 10    # number of traits
alpha_grid = sort(exp(seq(log(0.05), log(0.05/p), length.out=p*2)))    # alpha grid in HC statistic
nsim = 1e4    # number of MC iterations
null_dist <- TraitScan.MC(nsim = nsim, p = p, alpha_grid = alpha_grid)

beta_hat = c(rnorm(8,0,0.1), 0.3, 0.3)
se_hat = rep(0.1, p)
null_cov = matrix(0.2, nrow = p, ncol = p)
diag(null_cov) = 1

example <- TraitScan.summary.level(beta_hat, se_hat, null_cov, alpha_grid, null_dist)
}
\references{
Cao, Rui, et al. "Subset scanning for multi-trait analysis using GWAS summary statistics." medRxiv (2023): 2023-07.
}
\author{
Rui Cao, Tianzhong Yang
}
